refactor: Controller/Service를 Actor·URL 기준으로 그룹 병합 (REF-012)
기능 1개당 1클래스 구조를 Actor(Seller/Admin/WHM)·URL 리소스·공유 의존성
기준으로 묶어 파편화를 줄였다. CQRS 경계는 그대로 유지한다.

- Query Controller 9 → 4 (Seller/Admin/Whm/Dashboard)
- Command Controller 6 → 3 (SellerOrder/OrderStatus/BulkOrder)
- Seller Command Service 병합 (Create+Cancel → SellerOrderCommandService)
- @WebMvcTest 10개 대상/MockitoBean 재정렬
- 전체 118 테스트 GREEN
